wave 34: ~1150B data cleared (6 units) + 5 matches + sqrtf-ballast census (52 carriers, 6 cured)
itkusudama/gmtou/grzebes/mncharsel/tydisplay/itsamusgrapple byte-exact
or near; grZebes_801DA0C4 28->100; extab-coupled ballast caveat (J1);
epsilon-widened f64 compare law (J5); invented-vararg drops pool-proven.

+++ src/melee/gm/gm_16AE.c
@@ -286,7 +286,7 @@ void gm_8016B378(s8 arg0)
     lbl_8046B6A0.unk_18 = arg0;
 }
 
-void fn_8016B388(int arg0, s16 arg1)
+void fn_8016B388(int arg0, int arg1)
 {
     lbl_8046B6A0.FighterMatchInfo[arg0].x6 = arg1;
 }
